% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temperature_function.R
\name{convert_Fahrenheid_to_Celsius}
\alias{convert_Fahrenheid_to_Celsius}
\title{Fahrenheit conversion}
\usage{
convert_Fahrenheid_to_Celsius(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
